+++ src/components/Editor/Runners/HtmlRunner/HtmlRunner.js
@@ -1,28 +1,66 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import "./HtmlRunner.scss";
 import React, { useRef, useEffect } from "react";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
+import { codeRunHandled } from "../../EditorSlice";
 
 function HtmlRunner() {
   const projectCode = useSelector((state) => state.editor.project.components);
+  const focussedFileIndex = useSelector((state) => state.editor.focussedFileIndex);
+  const openFiles = useSelector((state) => state.editor.openFiles);
+  const codeRunTriggered = useSelector((state) => state.editor.codeRunTriggered)
+  const justLoaded = useSelector((state) => state.editor.justLoaded)
+
+  const dispatch = useDispatch()
   const output = useRef();
+
+
+  const htmlFiles = projectCode.filter(
+    (component) => component.extension === 'html'
+  )
+
+  const cssFiles = projectCode.filter(
+    (component) => component.extension === "css"
+  );
+
+  const fileName = openFiles[focussedFileIndex]
+  const focussedComponent = projectCode.find((component) => `${component.name}.${component.extension}` === fileName)
+
 
   const getBlobURL = (code, type) => {
     const blob = new Blob([code], { type });
     return URL.createObjectURL(blob);
   };
 
+  let timeout;
+
   useEffect(() => {
-    runCode();
-  }, [projectCode]);
+    if (justLoaded) {
+      runCode()
+    } else {
+      timeout = setTimeout(() => {
+        runCode()
+      }, 2000);
+      return () => clearTimeout(timeout)
+    }
+  }, [projectCode, focussedFileIndex]);
+
+  useEffect(() => {
+    if (codeRunTriggered) {
+      runCode()
+    }
+  }, [codeRunTriggered])
+
+  useEffect(() => {
+    runCode()
+  }, [focussedFileIndex])
 
   const runCode = () => {
     // TODO: get html files and handle urls for non index pages
-    let indexPage = projectCode[0].content;
+    const indexHTML = htmlFiles.find((component) => `${component.name}.${component.extension}` === 'index.html')
+    const componentToPreview = focussedComponent.extension === 'html' ? focussedComponent : indexHTML
+    let indexPage = componentToPreview.content;
 
-    const cssFiles = projectCode.filter(
-      (component) => component.extension === "css"
-    );
     cssFiles.forEach((cssFile) => {
       const cssFileBlob = getBlobURL(cssFile.content, "text/css");
       indexPage = indexPage.replace(
@@ -33,6 +71,10 @@ function HtmlRunner() {
 
     const blob = getBlobURL(indexPage, "text/html");
     output.current.src = blob;
+    if (codeRunTriggered) {
+      dispatch(codeRunHandled())
+    }
+    clearTimeout(timeout)
   };
 
   return (