+++ src/components/Modals/NewFileModal.js
@@ -59,7 +59,10 @@ const NewFileModal = () => {
       </div>
 
       <div className='modal-content__body'>
-        <label htmlFor='name'>{t('filePane.newFileModal.inputLabel')}</label>
+        <div>
+          <label htmlFor='name'>{t('filePane.newFileModal.inputLabel')}</label>
+          <p className='modal-content__help-text'>{t('filePane.newFileModal.helpText', {examples: t(`filePane.newFileModal.helpTextExample.${projectType}`)})}</p>
+        </div>
         <div>
           <NameErrorMessage />
           <input ref={nameInput} type='text' name='name' id='name' onKeyDown={onKeyDown}></input>
@@ -67,7 +70,7 @@ const NewFileModal = () => {
       </div>
 
       <div className='modal-content__buttons'>
-        <Button className='btn--primary' buttonText={t('filePane.newFileModal.save')} onClickHandler={createComponent} />
+        <Button className='btn--primary' buttonText={t('filePane.newFileModal.addFile')} onClickHandler={createComponent} />
         <Button className='btn--secondary' buttonText={t('filePane.newFileModal.cancel')} onClickHandler={closeModal} />
       </div>
     </Modal>